refactor(test): rename test/extend to test/extended
Rename test/extend directory to test/extended for better naming
convention and update build tags to use underscores instead of
hyphens (required by Go build tag syntax).

Changes:
- Rename directory test/extend -> test/extended
- Update build tag e2e_extended -> test_extended
- Update package name from extend to extended
- Add GO_EXTENDED_RECIPE variable in Makefile
- Update all import paths to use test/extended

The GO_EXTENDED_RECIPE follows the same pattern as other test
recipes (GO_E2E_RECIPE, GO_E2EV2_RECIPE) for consistency.
