+++ test/extended/util/hostedclusters.go
@@ -1,4 +1,4 @@
-package extend
+package util
 
 import (
 	"context"
