+++ test/extended/util/hypershift_cluster.go
@@ -1,4 +1,4 @@
-package extend
+package util
 
 import (
 	"context"
